+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   * dumped `refresh` stage config
   * changed `remove` command to leave the project as-is
   * dumped svn credentials handling
+  * dumped macros
   
 * stool is no longer responsible for building - use your standard build tools and configuration instead
   * dumped `build` command